Added Tree object
